Activity tab categories module design (#5833)
* - adds TopCategoriesView, separate loadCategories function and load function in onResume

* - adds full views for top categories view
- adds interaction, code fixes, ui updates and resource files

* - adds click action, navigation logics
- ui/code fixes

* - ui fix

* - code/design fixes

* Tiny bar chart.

* Yes.

---------

+++ app/src/main/java/org/wikipedia/history/db/HistoryEntryDao.kt
@@ -40,7 +40,7 @@ interface HistoryEntryDao {
     suspend fun deleteBy(authority: String, lang: String, namespace: String?, apiTitle: String)
 
     @Query("SELECT COUNT(*) FROM HistoryEntry WHERE timestamp > :timestamp")
-    suspend fun getTotalEntriesSince(timestamp: Long): Long?
+    suspend fun getTotalEntriesSince(timestamp: Long): Int?
 
     @Query("SELECT * FROM HistoryEntry ORDER BY timestamp DESC LIMIT 1")
     suspend fun getMostRecentEntry(): HistoryEntry?
